--- AgendamentoSistema/AgendamentoSistema.Models/Usuario.cs
@@ -1,7 +1,0 @@
-﻿namespace AgendamentoSistema.Models
-{
-    public class Usuario
-    {
-
-    }
-}
