+++ AgendamentoSistema/AgendamentoSistema.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 17.13.35931.197
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AgendamentoSistema", "AgendamentoSistema.csproj", "{1BC3DA9A-A483-39D9-8298-30DC93958751}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AgendamentoSistema.Models", "AgendamentoSistema.Models\AgendamentoSistema.Models.csproj", "{9B219ADF-FE98-4668-BC0B-5A59DA18DDD0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SistemaAgendamento.Models", "..\SistemaAgendamento.Models\SistemaAgendamento.Models.csproj", "{B5BDC0EC-17E5-4F02-8C19-01EA93A159DB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SistemaAgendamento.Service", "..\SistemaAgendamento.Service\SistemaAgendamento.Service.csproj", "{D6037B54-7C5A-4111-A0AD-570873199340}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +19,14 @@ Global
 		{1BC3DA9A-A483-39D9-8298-30DC93958751}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1BC3DA9A-A483-39D9-8298-30DC93958751}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1BC3DA9A-A483-39D9-8298-30DC93958751}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9B219ADF-FE98-4668-BC0B-5A59DA18DDD0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9B219ADF-FE98-4668-BC0B-5A59DA18DDD0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9B219ADF-FE98-4668-BC0B-5A59DA18DDD0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9B219ADF-FE98-4668-BC0B-5A59DA18DDD0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B5BDC0EC-17E5-4F02-8C19-01EA93A159DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5BDC0EC-17E5-4F02-8C19-01EA93A159DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5BDC0EC-17E5-4F02-8C19-01EA93A159DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5BDC0EC-17E5-4F02-8C19-01EA93A159DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D6037B54-7C5A-4111-A0AD-570873199340}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D6037B54-7C5A-4111-A0AD-570873199340}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D6037B54-7C5A-4111-A0AD-570873199340}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D6037B54-7C5A-4111-A0AD-570873199340}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
