+++ AgendamentoSistema/Controllers/ServicosController.cs
@@ -9,156 +9,7 @@ namespace AgendamentoSistema.Controllers
 {
     public class ServicosController : Controller
     {
-        private readonly ApplicationDbContext _context;
 
-        public ServicosController(ApplicationDbContext context)
-        {
-            _context = context;
-        }
-
-        // GET: Servicos
-        public async Task<IActionResult> Index()
-        {
-            var servicos = await _context.Servicos.ToListAsync();
-            return View(servicos);
-        }
-
-        // GET: Servicos/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var servico = await _context.Servicos
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (servico == null)
-            {
-                return NotFound();
-            }
-
-            return View(servico);
-        }
-
-        // GET: Servicos/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Servicos/Create
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Nome,Descricao,DuracaoMinutos,Preco")] Servico servico)
-        {
-            ModelState.Remove("Id");
-            ModelState.Remove("Agendamentos");
-
-            if (ModelState.IsValid)
-            {
-                _context.Add(servico);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(servico);
-        }
-
-        // GET: Servicos/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var servico = await _context.Servicos.FindAsync(id);
-            if (servico == null)
-            {
-                return NotFound();
-            }
-            return View(servico);
-        }
-
-        // POST: Servicos/Edit/5
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Nome,Descricao,DuracaoMinutos,Preco")] Servico servico)
-        {
-            if (id != servico.Id)
-            {
-                return NotFound();
-            }
-
-            ModelState.Remove("Agendamentos");
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(servico);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ServicoExists(servico.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(servico);
-        }
-
-        // GET: Servicos/Delete/5
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var servico = await _context.Servicos
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (servico == null)
-            {
-                return NotFound();
-            }
-
-            return View(servico);
-        }
-
-        // POST: Servicos/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var servico = await _context.Servicos.FindAsync(id);
-            if (servico != null)
-            {
-                var agendamentosAssociados = await _context.Agendamentos.AnyAsync(a => a.ServicoId == id);
-                if (agendamentosAssociados)
-                {
-                    ModelState.AddModelError(string.Empty, "Não é possível excluir este serviço pois existem agendamentos associados a ele.");
-                    return View(servico);
-                }
-
-                _context.Servicos.Remove(servico);
-                await _context.SaveChangesAsync();
-            }
-            
-            return RedirectToAction(nameof(Index));
-        }
-
-        private bool ServicoExists(int id)
-        {
-            return _context.Servicos.Any(e => e.Id == id);
-        }
     }
 }
