+++ AgendamentoSistema/Controllers/UsuarioController.cs
@@ -1,0 +1,7 @@
+﻿namespace AgendamentoSistema.Controllers
+{
+    public class UsuarioController
+    {
+        
+    }
+}
