+++ SistemaAgendamento.Service/IServive.cs
@@ -1,0 +1,7 @@
+﻿namespace SistemaAgendamento.Service
+{
+    public class IServive
+    {
+
+    }
+}
